Use contexts for subctl verify
Allow the user to specify contexts as arguments to subctl verify,
instead of kubeconfigs; this allows contexts to be loaded from the
KUBECONFIG environment variable, and with the help of the loading fix
in Shipyard, avoids errors when the default context is set in the
kubeconfigs.

To ease the transition, kubeconfigs are still supported.

Depends on: https://github.com/submariner-io/shipyard/pull/511
Fixes: #1130
Signed-off-by: Stephen Kitt <skitt@redhat.com>