Drop the Secret syncer
We don't currently need it, and it may be counter-productive. The
Secret syncer could be the cause of the issues we sometimes see with
disappearing secrets, and it is also be the source of
unstructured.Unstructured errors (see
https://github.com/submariner-io/submariner-operator/issues/1921 for
details).

This patch drops the syncer setup and cancellation code. If it is
restored in future (because we end up wanting to sync secrets, instead
of relying on tokens tied to object lifetimes), the Unstructured
errors should be addressed.

Signed-off-by: Stephen Kitt <skitt@redhat.com>